# docs/swagger/swagger.yaml
definitions:
...
  SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  internal_adapters_server_handler.BusinessProfileResponse:
    properties:
      address:
        type: string
      category:
        type: string
      description:
        type: string
      email:
        type: string
      isBusiness:
        type: boolean
      jid:
        type: string
      name:
        type: string
      website:
        type: string
    type: object
...
  internal_adapters_server_handler.CheckWhatsAppRequest:
    properties:
      phoneNumbers:
        items:
          type: string
        maxItems: 50
        minItems: 1
        type: array
    required:
    - phoneNumbers
    type: object
  internal_adapters_server_handler.CheckWhatsAppResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/internal_adapters_server_handler.CheckWhatsAppResult'
        type: array
      total:
        type: integer
    type: object
  internal_adapters_server_handler.CheckWhatsAppResult:
    properties:
      error:
        type: string
      isOnWhatsApp:
        type: boolean
      jid:
        type: string
      phoneNumber:
        type: string
    type: object
  internal_adapters_server_handler.ContactInfo:
    properties:
      isBusiness:
        type: boolean
      isMyContact:
        type: boolean
      isOnWhatsApp:
        type: boolean
      jid:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
      pushName:
        type: string
      shortName:
        type: string
    type: object
  internal_adapters_server_handler.GetProfilePictureResponse:
    properties:
      hasPicture:
        type: boolean
      jid:
        type: string
      pictureId:
        type: string
      pictureUrl:
        type: string
    type: object
  internal_adapters_server_handler.GetUserInfoRequest:
    properties:
      jids:
        items:
          type: string
        maxItems: 20
        minItems: 1
        type: array
    required:
    - jids
    type: object
  internal_adapters_server_handler.GetUserInfoResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/internal_adapters_server_handler.UserInfoResult'
        type: array
      total:
        type: integer
    type: object
  internal_adapters_server_handler.ListContactsResponse:
    properties:
      contacts:
        items:
          $ref: '#/definitions/internal_adapters_server_handler.ContactInfo'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  internal_adapters_server_handler.SyncContactsResponse:
    properties:
      message:
        type: string
      status:
        type: string
      syncedContacts:
        type: integer
      totalContacts:
        type: integer
    type: object
  internal_adapters_server_handler.UserInfoResult:
    properties:
      error:
        type: string
      isBusiness:
        type: boolean
      isOnWhatsApp:
        type: boolean
      jid:
        type: string
      name:
        type: string
      pictureId:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /sessions/{sessionId}/contacts:
    get:
      description: List contacts with pagination and filters
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: 'Limit (default: 50, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Offset (default: 0)'
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.ListContactsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: List contacts
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/all:
    get:
      description: Get all contacts without pagination
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_server_handler.ContactInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all contacts
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/avatar:
    get:
      description: Get profile picture of a contact
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Contact JID
        in: query
        name: jid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.GetProfilePictureResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profile picture
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/business:
    get:
      description: Get business profile of a contact
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Contact JID
        in: query
        name: jid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.BusinessProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Get business profile
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/check:
    post:
      consumes:
      - application/json
      description: Check if phone numbers are registered on WhatsApp
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone numbers to check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_server_handler.CheckWhatsAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.CheckWhatsAppResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Check WhatsApp numbers
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/detailed-info:
    post:
      consumes:
      - application/json
      description: Get detailed information about WhatsApp users
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: User JIDs to get detailed info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_server_handler.GetUserInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.GetUserInfoResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Get detailed user info (batch)
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/info:
    post:
      consumes:
      - application/json
      description: Get information about WhatsApp users
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: User JIDs to get info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_server_handler.GetUserInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.GetUserInfoResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user info
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/is-on-whatsapp:
    post:
      consumes:
      - application/json
      description: Check if multiple phone numbers are registered on WhatsApp
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone numbers to check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_server_handler.CheckWhatsAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.CheckWhatsAppResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Check if numbers are on WhatsApp (batch)
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/profile-picture-info:
    get:
      description: Get profile picture information of a contact
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Contact JID
        in: query
        name: jid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.GetProfilePictureResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profile picture info
      tags:
      - Contacts
  /sessions/{sessionId}/contacts/sync:
    post:
      description: Sync contacts from the device
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_server_handler.SyncContactsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Sync contacts
      tags:
      - Contacts
...
swagger: "2.0"